+++ docs/changelog.md
@@ -9,6 +9,10 @@ Added:
 * Decayed photons can be included in the HDF5 output file, if desired, via new optional flag `dump_decayed_photons`.
 * Tracking of laser depletion: new output variable `absorption` can be used to examine how much energy electrons and positrons have drawn from the laser pulse.
 
+Fixed:
+
+* Inconsistent use of units in stats block: filter bounds are now interpreted as being given in SI units.
+
 ## v1.3.5
 
 2024-03-08